+++ imod/mf6/package.py
@@ -11,7 +11,7 @@ import jinja2
 import numpy as np
 import xarray as xr
 import xugrid as xu
-from xarray.core.utils import is_scalar
+
 
 import imod
 from imod.mf6.auxiliary_variables import (
@@ -540,59 +540,7 @@ class Package(PackageBase, IPackage, abc.ABC):
             The package with part masked.
         """
 
-        masked = {}
-        if hasattr(self,"auxiliary_data_fields"):
-            remove_expanded_auxiliary_variables_from_dataset(self)
-        for var in self.dataset.data_vars.keys():
-            if self._skip_masking_variable(var, self.dataset[var]):
-                masked[var] = self.dataset[var]
-            else:
-                masked[var] = self._mask_spatial_var(var, mask)
-        if hasattr(self,"auxiliary_data_fields"):
-            expand_transient_auxiliary_variables(self)
-        return type(self)(**masked)
-
-    def _skip_masking_variable(self, var: str, da: GridDataArray)->bool: 
-        if self._skip_masking_dataarray(var) or len(da.dims) == 0 or set(da.coords).issubset(["layer"]):
-            return True
-        if is_scalar(da.values[()]):
-            return True
-        spatial_dims = ["x", "y", "mesh2d_nFaces", "layer"]
-        if not np.any( [coord in spatial_dims for coord in da.coords]):
-            return True
-        return False
-
-    def _mask_spatial_var(self, var: str, mask: GridDataArray)->GridDataArray:
-        da = self.dataset[var]
-        array_mask = self._adjust_mask_for_unlayered_data(da, mask)
-
-        if issubclass(da.dtype.type, numbers.Integral):
-            if var == "idomain":
-                return da.where(array_mask > 0, other=array_mask)
-            else:
-                return da.where(array_mask > 0, other=0)
-        elif issubclass(da.dtype.type, numbers.Real):
-            return da.where(array_mask > 0)
-        else:
-            raise TypeError(
-                f"Expected dtype float or integer. Received instead: {da.dtype}"
-            )
-
-    def _adjust_mask_for_unlayered_data(self, da: GridDataArray, mask: GridDataArray)->GridDataArray:
-        '''
-        Some arrays are not layered while the mask is layered (for example the
-        top array in dis or disv packaged). In that case we use the top layer of
-        the mask to perform the masking. If layer is not a dataset dimension,
-        but still a dataset coordinate, we limit the mask to the relevant layer
-        coordinate(s). 
-        '''
-        array_mask  = mask
-        if "layer" in da.coords and "layer" not in da.dims:
-            array_mask = mask.sel(layer=da.coords["layer"])        
-        if "layer" not in da.coords and "layer" in array_mask.coords:
-            array_mask = mask.isel(layer=0)
-
-        return array_mask              
+     
 
 
     def regrid_like(
